+++ build.gradle
@@ -48,9 +48,23 @@ gradlePlugin {
     plugins {
         plugins {
             id = 'net.wooga.plugins'
-            displayName = 'Wooga Base Gradle Plugin'
-            description = 'This plugin provides tasks and basic setup for other plugins'
+            displayName = 'wooga base gradle plugin'
+            description = 'this plugin provides tasks and basic setup for other plugins'
             implementationClass = 'wooga.gradle.plugins.PluginsPlugin'
+        }
+
+        localPlugins {
+            id = 'net.wooga.plugins-local'
+            displayName = 'wooga base gradle plugin'
+            description = 'this plugin provides tasks and basic setup for other plugins'
+            implementationClass = 'wooga.gradle.plugins.LocalPluginsPlugin'
+        }
+
+        privatePlugins {
+            id = 'net.wooga.plugins-private'
+            displayName = 'wooga base gradle plugin for private plugins'
+            description = 'this plugin provides tasks and basic setup for private plugins'
+            implementationClass = 'wooga.gradle.plugins.PrivatePluginsPlugin'
         }
     }
 }
